Fixed error of nonunique ids. Fixed rendering error. Seperated BigCats and SingleCat.
